requirements.txt
annotated-types==0.7.0
anyio==4.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.27.1
idna==3.10
jiter==0.8.2
openai==1.59.8
packaging==24.2
pydantic==2.10.5
pydantic_core==2.27.2
python-telegram-bot==20.3
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
